requirements.txt
python-dotenv
openai
ipykernel
langchain_openai
langchain
langchain-google-genai
Pillow
langgraph
pypdf
langchain_community
langgraph-cli[inmem]
langchain-tavily
certifi
requests
urllib3
#hf-agent-course
smolagents
duckduckgo_search
smolagents[gradio]
rank-bm25
plotly
geopandas
shapely
kaleido
kaleido
ipywidgets
jupyter
tqdm
markdownify
helium
selenium
smolagents[all]
